splitting out hdf5 tests, descoping some test fixtures. trying to fix buggy travis tests
